Fix bug with including go in path
On debian, we use the `gophers/archive` ppa to install go. That repo
installs the go binary to `/usr/lib/go-{{ version }}/bin/go`. We do not
automatically include this directory in the path. Previously, we updated
the ~/.bashrc to include it only when we installed the package. However,
each `sheepdoge` run destructs the `~/.bashrc`, so we actually need to
update the ~/.bashrc everytime.